Updated repo for Percona-Xtrabackup (#4)
The current version that is included in this Dockerfile does not support the current release versions. Thus, I updated the Dockerfile to add the official repository from Percona and added a few additional commands to ensure no additional packages are installed as well as a cleanup process at the end.